Change file test health check Relate to #503
