+++ src/server/bootstrap.test.ts
@@ -1,0 +1,11 @@
+import axios from 'axios';
+import { assert } from 'chai';
+
+describe('health check test case', healthCheckTestCases);
+
+function healthCheckTestCases() {
+  it('should return empty array for now startTimestamp', async () => {
+    const result = await axios.get('http://localhost:4000/health');
+    assert.equal(result.data, 'Hi every thing seems ok');
+  });
+}
